+++ tests/postcss.test.ts
@@ -1,4 +1,5 @@
 import { assertSiteSnapshot, build, getSite } from "./utils.ts";
+import { assert } from "../deps/assert.ts";
 import postcss from "../plugins/postcss.ts";
 import nano from "npm:cssnano";
 
@@ -37,4 +38,30 @@ Deno.test("postcss plugin with hooks", async (t) => {
 
   await build(site);
   await assertSiteSnapshot(t, site);
+});
+
+Deno.test("postcss plugin with default name", () => {
+  const site = getSite({
+    src: "postcss",
+  });
+
+  site.use(postcss());
+
+  const { helpers } = site.renderer;
+
+  assert(helpers.has("postcss"));
+});
+
+Deno.test("postcss plugin with custom name", () => {
+  const site = getSite({
+    src: "postcss",
+  });
+
+  site.use(postcss({
+    name: "css",
+  }));
+
+  const { helpers } = site.renderer;
+
+  assert(helpers.has("css"));
 });